task-6: rename notification service file name
